pyproject.toml
[project]
name = "kingshot-heroes"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.120.4",
    "httpx>=0.27.0",
    "numpy>=2.3.4",
    "pandas>=2.3.3",
    "Pillow>=11.0.0",
    "pyarrow>=22.0.0",
    "pydantic>=2.12.3",
    "pytest>=8.4.2",
    "python-dotenv>=1.2.1",
    "python-jose[cryptography]>=3.3.0",
    "supabase>=2.23.0",
    "thefuzz>=0.22.1",
    "tqdm>=4.67.1",
    "uvicorn>=0.38.0",
]

[tool.pytest.ini_options]
pythonpath = ["src"]
addopts = ["--ignore=simulation"]
